fix(model-parsing): remove circular dependency
Pre-commit hook ran. Total eslint: 0, total circular: 1
